requirements.txt
uvicorn
fastapi
sqlmodel
dynaconf
loguru
psycopg2-binary
requests
google-genai
pycountry
pillow
